+++ src/CurrentProcess.php
@@ -8,6 +8,7 @@ use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
 use Innmind\TimeContinuum\Period;
 use Innmind\TimeWarp\Halt;
+use Innmind\Signals\Handler;
 use Innmind\Immutable\{
     Attempt,
     SideEffect,
@@ -16,19 +17,21 @@ use Innmind\Immutable\{
 final class CurrentProcess
 {
     private Halt $halt;
+    private Handler $handler;
     private ?Signals $signals = null;
 
-    private function __construct(Halt $halt)
+    private function __construct(Halt $halt, Handler $handler)
     {
         $this->halt = $halt;
+        $this->handler = $handler;
     }
 
     /**
      * @internal
      */
-    public static function of(Halt $halt): self
+    public static function of(Halt $halt, Handler $handler): self
     {
-        return new self($halt);
+        return new self($halt, $handler);
     }
 
     /**
@@ -47,7 +50,7 @@ final class CurrentProcess
 
     public function signals(): Signals
     {
-        return $this->signals ??= Signals::of();
+        return $this->signals ??= Signals::of($this->handler);
     }
 
     /**